src/workflows/test_receive_addresses/requirements.txt
--index-url https://nexus.engageska-portugal.pt/repository/pypi/simple
--extra-index-url https://pypi.org/simple
ska-logging
ska-sdp-workflow >= 0.1.0
